// src/types/Employee.ts
export default interface IEmployee {
  email: string;
  gitHub: string;
  highlighted: boolean;
  imagePortraitUrl: string;
  imageWallOfLeetUrl: string;
  linkedIn: string;
  mainText: string;
  manager: string;
  name: string;
  office: string;
  orgUnit: string;
  phoneNumber: string;
  published: boolean;
  stackOverflow: string | null;
  twitter: string;
}